Fix some hardcoded shard locations
If the shard goal is below 5 or 7, these fail to place and cause the fuzzer to fail that generation.
I attempted to mitigate that by instead changing the location to be dynamic based on the shard goal (around halfway and 75% of the way to the goal shard, respectively) With a shard goal of 10 these locations will be in the same shards as before.
(Variable names are hard 😆)